reducing use lazy loading to reduce initial js
