+++ src/components/link.tsx
@@ -3,7 +3,7 @@ import { ComponentProps } from "react"
 import { twMerge } from "tailwind-merge"
 
 export const buttonStyling =
-  "text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-primary-600 dark:hover:bg-primary-700 focus:outline-none dark:focus:ring-primary-800"
+  "inline-block text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-[1.125rem] py-2.5 dark:bg-primary-600 dark:hover:bg-primary-700 focus:outline-none dark:focus:ring-primary-800 border border-transparent cursor-pointer"
 
 export const Link = (props: ComponentProps<typeof NextLink>) => (
   <NextLink {...props} className={twMerge(props.className, buttonStyling)}></NextLink>
