--- src/lib/pdfviewer/index.d.ts
@@ -1,1 +1,0 @@
-export { PDFViewer, PDFDownloadLink } from "@react-pdf/renderer"
