--- src/lib/pdfviewer/index.js
@@ -1,3 +1,0 @@
-export { PDFViewer, PDFDownloadLink } from "@react-pdf/renderer/lib/react-pdf.browser.cjs.min.js"
-
-// to get around problems with PDFViewer and Nextjs SSR
