+++ src/pages/api/email.ts
@@ -18,7 +18,8 @@ import { config } from "@/lib/util/config"
 import { getUserData, storageBucket, user } from "@/lib/db"
 import { getDonations } from "@/lib/qbo-api"
 import { formatEmailBody } from "@/lib/email"
-import { WithBody, ReceiptPdfDocument } from "@/components/receipt"
+import { ReceiptPdfDocument } from "@/components/receipt/pdf"
+import { WithBody } from "@/components/receipt/email"
 import { Donation } from "@/types/qbo-api"
 import { EmailHistoryItem } from "@/types/db"
